Bumps version to 6.0.28 Official Release
I just patched a few things. Found out by the V8 port being down on FreeBSD, that V8 is not necessary, so I went ahead and got rid of Python27 for Python37... I also got rid of line 99, which is currently line 100 of the commit, this line for the base.txz file is completely unnecessary and does not effect the installation of the controller. I added a unlock and lock of the packages that are installed by the install-unifi.sh script... This was done for situations where a user might be running something like OPNSense where updates from OPNSense like to reinstall packages which can effectively break the UniFi controller. This currently is the most stable script yet, of course all depedency errors were also fixed.